+++ src/api/sk_sched.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-#include "flist/flist.h"
+#include "flibs/flist.h"
 #include "api/sk_txn.h"
 #include "api/sk_entity.h"
 #include "api/sk_entity_mgr.h"