+++ src/user-c/src/logger_private.c
@@ -1,11 +1,11 @@
 #include <stdlib.h>
 
-#include "flog/flog.h"
+#include "flibs/flog.h"
 #include "api/sk_log.h"
 #include "api/sk_log_tpl.h"
 #include "api/sk_env.h"
 
-#include "skull/skull_api.h"
+#include "skull/logger_private.h"
 
 void skull_log(const char* fmt, ...)
 {